completed deliverables
