+++ sql_commands.txt
@@ -1,0 +1,10 @@
+Figure out the appropriate SQL commands to find out the following, and keep track of them in the file called `sql_commands.txt`
+
+1. All columns for all players from the New York Knicks (NYK)
+    SELECT from nba_db WHERE team = 'NYK';
+
+1. All columns for all players from the Indiana Packers (IND) who are under 26 years old
+    SELECT from nba_db WHERE team = 'NYK' AND age < 26;
+
+1. All columns for all players, ordered from least points scored to most points scored
+    SELECT from nba_db ORDER BY points ASC;
